// packages/plural-api/src/types/enums.ts
export enum SocketOperationType {
  Insert = 'insert',
  Update = 'update',
  Delete = 'delete',
}

export enum PluralVisibility {
  Trusted = 'Trusted',
  Private = 'Private',
  Public = 'Public',
}

export enum MemberFieldType {
  String = 'String',
  Color = 'Color',
  Date = 'Date',
  Month = 'Month',
  Year = 'Year',
  MonthYear = 'MonthYear',
  Timestamp = 'Timestamp',
  MonthDay = 'MonthDay',
}